Add auth check for update and delete

+++ controllers/feed.js
@@ -128,6 +128,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error; // throw to the catch block which uses next()
       }
+      // Checking if logged in user owns the post.
+      if (post.creator.toString() !== req.userId.toString()) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -158,6 +165,12 @@ exports.deletePost = (req, res, next) => {
         const error = new Error("Could not find the post.");
         error.statusCode = 404;
         throw error; // throw to the catch block which uses next()
+      }
+      // Checking if logged in user owns the post.
+      if (post.creator.toString() !== req.userId.toString()) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
       }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
